requirements.txt
# Core Dependencies
endee==0.1.8
openai==1.12.0
groq>=1.0.0
python-dotenv==1.2.1
numpy>=2.2.4

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
requests==2.31.0

# Web Interface
streamlit==1.31.0

# Visualization and Evaluation
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
scikit-learn>=1.5.0

# Optional: For enhanced functionality
langchain==0.1.6
sentence-transformers==2.3.1
tiktoken==0.5.2

# Development
pytest==7.4.4
black==24.1.1
